vendor: Generate OTA JSON for all official builds

+++ scripts/update.sh
@@ -3,9 +3,14 @@
 zip=$(get_build_var XTENDED_VERSION).zip
 device=$(echo $TARGET_PRODUCT | cut -d '_' -f2)
 buildprop=$OUT/system/build.prop
-romtype="OFFICIAL"
+linen=$(grep -n "ro.xtended.build.type" $buildprop | cut -d ':' -f1)
+romtype=$(sed -n $linen'p' < $buildprop | cut -d '=' -f2)
 variant=$(echo $zip | cut -d '-' -f4)
 name=$device'_'$variant.json
+
+if [ "$romtype" != "OFFICIAL" ]; then
+	return 0
+fi
 
 if [ -f $OUT/$name ]; then
 	rm $OUT/$name
